+++ internal/key/key.go
@@ -6,27 +6,20 @@ package key
 
 import (
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
 	"github.com/ava-labs/avalanchego/vms/platformvm"
 )
 
 // Key defines methods for key manager interface.
 type Key interface {
-	Addresser
-	Spender
-}
-
-type Addresser interface {
 	// P returns the formatted P-Chain address.
 	P() string
-
 	// Address returns the raw ids.ShortID address.
 	Address() ids.ShortID
-}
-
-type Spender interface {
 	// Spend attempts to spend all specified UTXOs (outputs)
 	// and returns the new UTXO inputs.
+	//
 	// If target amount is specified, it only uses the
 	// outputs until the total spending is below the target
 	// amount.
@@ -34,9 +27,8 @@ type Spender interface {
 		totalBalanceToSpend uint64,
 		inputs []*avax.TransferableInput,
 	)
-
-	// Sign applies [sigs] signatures to the provied [pTx].
-	Sign(pTx *platformvm.Tx, sigs int) error
+	// Sign generates [numSigs] signatures and attaches them to [pTx].
+	Sign(pTx *platformvm.Tx, numSigs int) error
 }
 
 type Op struct {
@@ -70,5 +62,18 @@ func WithTargetAmount(ta uint64) OpOption {
 func WithFeeDeduct(fee uint64) OpOption {
 	return func(op *Op) {
 		op.feeDeduct = fee
+	}
+}
+
+func getHRP(networkID uint32) string {
+	switch networkID {
+	case constants.LocalID:
+		return constants.LocalHRP
+	case constants.FujiID:
+		return constants.FujiHRP
+	case constants.MainnetID:
+		return constants.MainnetHRP
+	default:
+		return constants.FallbackHRP
 	}
 }
